-Use ChaosShell in chaos.exe rather than System.CommandLine -Use App.Current instead of Application.Current -Implement support for running multiple WPF unit tests utilizing the App at once -Cordb improvements -DeferrableOperation fixes -NativeReflector enhancements -Miscellaneous improvements

+++ src/ChaosDbg.Tests/Services/CordbEngine/CordbEngineLaunchTests.cs
@@ -2,8 +2,11 @@
 using System.Diagnostics;
 using System.Threading;
 using ChaosDbg.Cordb;
+using ChaosDbg.Metadata;
+using ChaosLib;
 using ClrDebug;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using TestApp;
 
 namespace ChaosDbg.Tests
 {
@@ -16,16 +19,115 @@ namespace ChaosDbg.Tests
             var thread = new Thread(() =>
             {
                 var cordebug = new CorDebug();
+                CordbLauncher.LogNativeCorDebugInfo(cordebug);
                 var cb = new CorDebugManagedCallback();
+
+                var exited = new ManualResetEventSlim(false);
+                cb.OnDebuggerError += (s, e) => exited.Set();
+                cb.OnExitProcess += (s, e) => exited.Set();
                 cordebug.SetManagedHandler(cb);
 
-                cordebug.CreateProcess("powershell.exe");
+                var process = cordebug.CreateProcess("powershell.exe");
+                CordbLauncher.LogNativeCorDebugProcessInfo(process);
+
+                process.Stop(0);
+                process.Terminate(0);
+
+                exited.Wait();
+
+                cordebug.Terminate();
             });
+            Log.CopyContextTo(thread);
 
             thread.SetApartmentState(ApartmentState.MTA);
 
             thread.Start();
             thread.Join();
+        }
+
+        [TestMethod]
+        public void CordbEngine_Launch_Crash_AfterStartup_OnManagedEventThread_ManagedOnly()
+        {
+            void DoTest()
+            {
+                TestSignalledDebugCreate(
+                    TestType.CordbEngine_Thread_StackTrace_ManagedFrames,
+                    action: ctx => ctx.WaitForFatalShutdown(),
+                    hookEvents: provider =>
+                    {
+                        provider.ModuleLoad += (s, e) => throw new DummyException();
+                    },
+                    waitForSignal: false
+                );
+            }
+
+            AssertEx.Throws<DummyException>(DoTest, "DummyException");
+        }
+
+        [TestMethod]
+        public void CordbEngine_Launch_Crash_AfterStartup_OnManagedEventThread_Interop()
+        {
+            void DoTest()
+            {
+                TestSignalledDebugCreate(
+                    TestType.CordbEngine_Thread_StackTrace_ManagedFrames,
+                    action: ctx => ctx.WaitForFatalShutdown(),
+                    hookEvents: provider =>
+                    {
+                        provider.ModuleLoad += (s, e) =>
+                        {
+                            if (e.Module is not CordbNativeModule)
+                                throw new DummyException();
+                        };
+                    },
+                    useInterop: true,
+                    handleLoaderBP: true,
+                    waitForSignal: false
+                );
+            }
+
+            AssertEx.Throws<DummyException>(DoTest, "DummyException");
+        }
+
+        [TestMethod]
+        public void CordbEngine_Launch_Crash_AfterStartup_OnUnmanagedEventThread_OutOfBand_Interop() =>
+            TestUnmanagedEventThreadCrash(crashInOutOfBand: true);
+
+        private void TestUnmanagedEventThreadCrash(bool crashInOutOfBand)
+        {
+            void DoTest()
+            {
+                TestSignalledDebugCreate(
+                    TestType.CordbEngine_Thread_StackTrace_ManagedFrames,
+                    action: ctx => ctx.WaitForFatalShutdown(),
+                    hookEvents: provider =>
+                    {
+                        provider.ModuleLoad += (s, e) =>
+                        {
+                            if (e.Module is CordbNativeModule m)
+                            {
+                                var isOutOfBand = m.Process.Session.CallbackContext.UnmanagedOutOfBand;
+
+                                if (crashInOutOfBand)
+                                {
+                                    if (isOutOfBand)
+                                        throw new DummyException();
+                                }
+                                else
+                                {
+                                    if (!isOutOfBand)
+                                        throw new DummyException();
+                                }
+                            }
+                        };
+                    },
+                    useInterop: true,
+                    handleLoaderBP: true,
+                    waitForSignal: false
+                );
+            }
+
+            AssertEx.Throws<DummyException>(DoTest, "DummyException");
         }
 
         [TestMethod]
@@ -36,8 +138,8 @@ namespace ChaosDbg.Tests
                 : "C:\\Windows\\SysWOW64\\WindowsPowerShell\\v1.0\\powershell.exe";
 
             AssertEx.Throws<DebuggerInitializationException>(
-                () => NetFrameworkProcess.Create(new CreateProcessOptions(path), null),
-                "process most likely does not match architecture of debugger"
+                () => CordbLauncher.Create(new CreateProcessTargetOptions(path) { FrameworkKind = FrameworkKind.NetCore }, null, default),
+                "however debugger process is"
             );
         }
 
@@ -45,7 +147,7 @@ namespace ChaosDbg.Tests
         public void CordbEngine_NetCore_Create_WrongArchitecture()
         {
             AssertEx.Throws<DebuggerInitializationException>(
-                () => NetCoreProcess.Create(new CreateProcessOptions(GetTestAppPath(matchCurrentProcess: false, netCore: true)), null),
+                () => CordbLauncher.Create(new CreateProcessTargetOptions(GetTestAppPath(matchCurrentProcess: false, netCore: true)) { FrameworkKind = FrameworkKind.NetCore }, null, default),
                 "however debugger process is"
             );
         }
@@ -66,8 +168,8 @@ namespace ChaosDbg.Tests
             try
             {
                 AssertEx.Throws<DebuggerInitializationException>(
-                    () => NetFrameworkProcess.Attach(new AttachProcessOptions(process.Id), null),
-                    "process most likely does not match architecture of debugger"
+                    () => CordbLauncher.Attach(new AttachProcessTargetOptions(process.Id) { FrameworkKind = FrameworkKind.NetFramework }, null, default),
+                    "however debugger process is"
                 );
             }
             finally
@@ -82,7 +184,7 @@ namespace ChaosDbg.Tests
         {
             var path = GetTestAppPath(matchCurrentProcess: false, netCore: true);
 
-            var psi = new ProcessStartInfo(path, EventName)
+            var psi = new ProcessStartInfo(path, $"{TestType.CordbEngine_Thread_StackTrace_ManagedFrames} {EventName}")
             {
                 WindowStyle = ProcessWindowStyle.Minimized
             };
@@ -92,13 +194,14 @@ namespace ChaosDbg.Tests
             try
             {
                 AssertEx.Throws<DebuggerInitializationException>(
-                    () => NetCoreProcess.Attach(new AttachProcessOptions(process.Id), null),
+                    () => CordbLauncher.Attach(new AttachProcessTargetOptions(process.Id) { FrameworkKind = FrameworkKind.NetCore }, null, default),
                     "however debugger process is"
                 );
             }
             finally
             {
-                process.Kill();
+                if (!process.HasExited)
+                    process.Kill();
             }
         }
     }
